BP-12 | moving send-job to hearing_map
